+++ packages/express/src/index.ts
@@ -1,3 +1,5 @@
+import { URL } from 'url';
+
 import { milliSecondTimer, sendApilyticsMetrics } from '@apilytics/core';
 import type { NextFunction, Request, RequestHandler, Response } from 'express';
 
@@ -15,9 +17,14 @@ export const apilyticsMiddleware = (
   return (req: Request, res: Response, next: NextFunction): void => {
     const timer = milliSecondTimer();
     res.on('finish', () => {
+      const { pathname: path, search: query } = new URL(
+        req.originalUrl,
+        'http://_', // Cannot parse a relative URL, so make it absolute.
+      );
       sendApilyticsMetrics({
         apiKey,
-        path: req.path,
+        path,
+        query,
         method: req.method,
         statusCode: res.statusCode,
         timeMillis: timer(),