+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Send query parameters in addition to the path.
+
 ## [1.0.2] - 2022-01-16
 
 ### Added